requirements.txt
# Core Dependencies
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Data Processing
ta-lib>=0.4.28
pandas-ta>=0.3.14b0
statsmodels>=0.14.0

# Web Framework
flask>=2.3.0
flask-sqlalchemy>=3.0.5
flask-login>=0.6.2
flask-wtf>=1.1.1

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
alembic>=1.11.0

# Visualization
plotly>=5.15.0
dash>=2.11.0
dash-bootstrap-components>=1.4.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Development Tools
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
isort>=5.12.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
schedule>=1.2.0

# Machine Learning
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.0

# Time Series Analysis
arch>=5.4.0
pmdarima>=2.0.3

# Email Notifications
flask-mail>=0.9.1 